Compile semantic css with esbuild
Instead of a bunch of very complicated loaders in webpack, let's just
use esbuild
