+++ apps/client/assets/css/index.less
@@ -1,4 +1,4 @@
-/* This file is for your main application css. */
+@import "semantic-ui-less/semantic.less";
 
 @background-color: black;
 @brand-primary: #AC17C6; // rgb(172, 23, 198)
