+++ apps/client/assets/package.json
@@ -3,7 +3,7 @@
   "license": "MIT",
   "scripts": {
     "bundle:static-css": "npx postcss ./static-css/index.scss -o ./../priv/static/static-css/index.css",
-    "bundle:static": "node ./scripts/compile_static.js",
+    "bundle:static": "node ./scripts/compile_static.mjs",
     "bundle:react": "NODE_ENV=production node ./node_modules/.bin/webpack --config webpack.config.production.js",
     "start": "node ./node_modules/.bin/webpack serve --config webpack.config.development.js",
     "s": "yarn start",
@@ -58,6 +58,7 @@
     "cssnano": "^4.1.11",
     "date-fns": "^2.24.0",
     "esbuild": "^0.13.3",
+    "esbuild-plugin-less": "^1.1.4",
     "eslint": "^7.32.0",
     "eslint-config-prettier": "^8.3.0",
     "eslint-plugin-prettier": "^4.0.0",
