+++ apps/client/assets/semantic-theme/theme.config
@@ -78,9 +78,10 @@
          Import Theme
 *******************************/
 
-@import (multiple) "~semantic-ui-less/theme.less";
+@import (multiple) "semantic-ui-less/theme.less";
 /* @fontPath "../../semantic-theme/site"; */
 
-@fontPath : "../../../themes/@{theme}/assets/fonts";
+@fontPath: "../node_modules/semantic-ui-less/themes/default/assets/fonts";
+@imagePath: "../node_modules/semantic-ui-less/themes/default/assets/images";
 
 /* End Config */
