+++ apps/client/assets/static/index.ejs
@@ -3,6 +3,7 @@
     <title><%= title %></title>
     <meta name="viewport" content="width=device-width, initial-scale=1">
     <link rel="stylesheet" href="/static-css/index.css">
+    <link rel="stylesheet" href="/css/index.css">
   </head>
   <div id="wee" data-https="<%= isHttps %>" />
 </html>
