+++ apps/client/assets/webpack.config.development.js
@@ -40,39 +40,6 @@ const webpackConfig = {
         include: path.resolve(__dirname, "js"),
         exclude: /node_modules/,
       },
-      // Handle js and jsx
-      //{
-      //test: /\.(js|jsx)$/,
-      //use: "babel-loader",
-      //include: path.resolve(__dirname, "js"),
-      //exclude: /node_modules/
-      //},
-      // Handle semantic-ui images
-      {
-        test: /\.jpe?g$|\.gif$|\.png$|\.ttf$|\.eot$|\.svg$/,
-        loader: "file-loader",
-        options: {
-          name: "[name].[ext]?[hash]",
-          outputPath: "fonts/",
-          publicPath: "../",
-        },
-        include: [path.resolve(__dirname, "node_modules/semantic-ui-less")],
-      },
-      // Handle semantic-ui fonts
-      {
-        test: /|\.woff$|\.woff2/,
-        loader: "file-loader",
-        options: {
-          name: "[name].[ext]?[hash]",
-          publicPath: "../",
-        },
-        include: [
-          path.resolve(
-            __dirname,
-            "node_modules/semantic-ui-less/themes/default/assets/fonts"
-          ),
-        ],
-      },
       // Handle static files
       {
         test: /\.(pdf|docx)$/,
@@ -81,15 +48,6 @@ const webpackConfig = {
           name: "[name].[ext]?[hash]",
         },
         include: path.resolve(__dirname, "static/files"),
-      },
-      // Handle less (semantic-ui + ours)
-      {
-        test: /\.less$/,
-        use: [MiniCssExtractPlugin.loader, "css-loader", "less-loader"],
-        include: [
-          path.resolve(__dirname, "css"),
-          path.resolve(__dirname, "node_modules/semantic-ui-less"),
-        ],
       },
     ],
   },
@@ -111,10 +69,6 @@ const webpackConfig = {
 
   resolve: {
     alias: {
-      "../../theme.config$": path.join(
-        __dirname,
-        "semantic-theme/theme.config"
-      ),
       "@store$": path.resolve(__dirname, "js/store/store"),
       "@store": path.resolve(__dirname, "js/store/"),
       "@utils": path.resolve(__dirname, "js/utils/"),
